[feature #162171018] update from with signup validation branch
