+++ .eslintrc
@@ -1,6 +1,5 @@
 {
   "root": true,
-  "parser": "babel-eslint" ,
   "extends": "airbnb-base",
   "env": {
     "node": true,
@@ -12,24 +11,30 @@
     "ecmaVersion": 6
   },
   "rules": {
-    "consistent-return":"error",
+    "consistent-return": "error",
     "no-param-reassign": "warn",
     "comma-dangle": 0,
     "curly": ["error", "multi-line"],
     "import/no-unresolved": [2, { "commonjs": true }],
     "no-shadow": ["error", { "allow": ["req", "res", "err"] }],
-    "valid-jsdoc": ["error", {
-      "requireReturn": true,
-      "requireReturnType": true,
-      "requireParamDescription": false,
-      "requireReturnDescription": true
-    }],
-    "require-jsdoc": ["error", {
+    "valid-jsdoc": [
+      "error",
+      {
+        "requireReturn": true,
+        "requireReturnType": true,
+        "requireParamDescription": false,
+        "requireReturnDescription": true
+      }
+    ],
+    "require-jsdoc": [
+      "error",
+      {
         "require": {
-            "FunctionDeclaration": true,
-            "MethodDefinition": true,
-            "ClassDeclaration": true
+          "FunctionDeclaration": true,
+          "MethodDefinition": true,
+          "ClassDeclaration": true
         }
-    }]
+      }
+    ]
   }
 }
