+++ routes/Users.js
@@ -5,9 +5,8 @@ import { passwordReset, validEmail } from '../middlewares/passwordReset';
 
 const router = express.Router();
 
-router.post('/forgotpassword', validEmail, userController.forgotPassword);
+router.post('/', validEmail, userController.forgotPassword);
 router.put('/resetpassword/:token', passwordReset, userController.resetPassword);
-
 
 router.use(bodyParser.json());
 router.use(bodyParser.urlencoded({ extended: false }));
