+++ routes/index.js
@@ -1,3 +1,4 @@
 import auth from './auth';
+import user from './users';
 
-export default auth;
+export { auth, user };
